requirements.txt
python-telegram-bot>=21,<22
python-dotenv>=1.0
pytz>=2024.1